programacion/requirements.txt
mysql-connector-python # mysql conector con python
configparser
colorama
pytest
# pip install -r requirements.txt
